WRR-12625: Adopted ESLint flat config (#365)
* updated eslint to version 9 and created new eslint config file

* linked eslint-config-enact branch that supports flat config

* removed unnecessary files property

* updated CHANGELOG.md

* adjusted change log

* updated eslint to latest minor version

* removed shrinkwrap

* added npm-shrinkwrap.json

* refactored eslint.config.js file to avoid plugin duplication

* replaced --no-eslintrc with --no-config-lookup

* updated config for eslint-webpack-plugin

* created separate eslint config for eslint-webpack-plugin

* updated shrinkwrap

* fixed shrinkwrap

* eslint fixes

* removed unused parameter

---------
