+++ eslint.config.js
@@ -1,0 +1,47 @@
+const enactConfig = require('eslint-config-enact');
+const prettierConfig = require('eslint-config-prettier');
+const importPlugin = require('eslint-plugin-import');
+const prettierPlugin = require('eslint-plugin-prettier');
+const globals = require('globals');
+
+module.exports = [
+	...enactConfig,
+	{
+		languageOptions: {
+			ecmaVersion: 'latest',
+			sourceType: 'module',
+			globals: {
+				...globals.node
+			},
+			parserOptions: {
+				ecmaFeatures: {
+					jsx: true
+				}
+			}
+		},
+		plugins: {
+			import: importPlugin,
+			prettier: prettierPlugin
+		},
+		rules: {
+			// import plugin rules
+			'import/no-unresolved': ['error', {commonjs: true, caseSensitive: true}],
+			'import/named': 'error',
+			'import/first': 'warn',
+			'import/no-duplicates': 'error',
+			'import/extensions': ['warn', 'always', {js: 'never', json: 'always'}],
+			'import/newline-after-import': 'warn',
+			'import/order': [
+				'warn',
+				{
+					'newlines-between': 'never',
+					groups: ['builtin', 'external', 'internal', 'parent', 'sibling', 'index']
+				}
+			],
+
+			// prettier rules
+			...prettierPlugin.configs.recommended.rules,
+			...prettierConfig.rules
+		}
+	}
+];
